Move varied_offset to builders/EUDEMO/tools
This was the suggested place for the function to live. My mistake
creating a new file in tools/first_wall
